requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
openai==1.53.0
httpx==0.27.2
werkzeug==3.0.3
itsdangerous==2.2.0
email-validator==2.2.0
